pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["WrldEngine <kamscienceit123.kamran@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-cleanup = "^8.1.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
black = "^24.4.0"
celery = "^5.4.0"
redis = "^5.0.3"
six = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
